split out `rust_ir` module
