+++ chalk-macros/src/lib.rs
@@ -1,5 +1,4 @@
 #![feature(crate_visibility_modifier)]
-#![feature(macro_vis_matcher)]
 #![feature(non_modrs_mods)]
 
 use std::cell::RefCell;
