+++ src/coherence/orphan.rs
@@ -2,6 +2,7 @@ use std::sync::Arc;
 
 use errors::*;
 use ir::*;
+use rust_ir::*;
 use cast::*;
 use solve::SolverChoice;
