+++ src/lib.rs
@@ -34,6 +34,8 @@ crate mod zip;
 #[macro_use]
 pub mod ir;
 
+pub mod rust_ir;
+
 crate mod cast;
 crate mod coherence;
 crate mod rules;
