+++ src/zip.rs
@@ -1,6 +1,7 @@
 use fallible::*;
 use fold::Fold;
 use ir::*;
+use rust_ir::*;
 use std::fmt::Debug;
 use std::sync::Arc;
